requirements.txt
asgiref==3.7.2
base58==2.1.1
bitcoin==1.1.39
blockcypher==1.0.93
cachetools==5.3.1
certifi==2023.7.22
cffi==1.15.1
charset-normalizer==3.2.0
cloudinary==1.34.0
cryptography==41.0.3
defusedxml==0.7.1
diff-match-patch==20230430
dj-rest-auth==5.0.0
Django==3.2.21
django-allauth==0.56.1
django-cloudinary-storage==0.3.0
django-cors-headers==4.1.0
django-environ==0.11.2
django-filter==23.2
django-import-export==3.2.0
djangorestframework==3.14.0
djangorestframework-simplejwt==5.3.0
ecdsa==0.18.0
et-xmlfile==1.1.0
gunicorn==21.2.0
idna==3.4
MarkupPy==1.14
oauthlib==3.2.2
odfpy==1.4.1
openpyxl==3.1.2
packaging==23.1
Pillow==9.5.0
psycopg2==2.9.6
pycparser==2.21
PyJWT==2.8.0
python-dateutil==2.8.2
python-dotenv==1.0.0
python3-openid==3.2.0
pytz==2023.3.post1
PyYAML==6.0.1
requests==2.31.0
requests-oauthlib==1.3.1
six==1.16.0
sqlparse==0.4.4
tablib==3.4.0
urllib3==1.26.16
whitenoise==6.5.0
xlrd==2.0.1
xlwt==1.3.0
